BACKPORT: staging/android: remove driver_data from struct sync_fence_info
It is unclear in what situations driver_data should be used thus better do
not upstream it for now. If a need arises in the future a discussion can
be started to re-add it.

Signed-off-by: Gustavo Padovan <gustavo.padovan@collabora.co.uk>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit c7434b8436f9fb79855f0a52e4e9665becb77aea)
from iwlwifi-next-for-kalle-2016-03-30
Signed-off-by: Tarun <tarun.vyas@intel.com>

BUG=b:36446508
TEST=No regressions in KBL and APL based devices

 Conflicts: [Tarun: "41576bbd CHROMIUM: android: fix warning when
		 releasing active sync point" added
		 android_fence_disable_signaling() function. Retain it]
	drivers/staging/android/sync.c

Change-Id: I961a12d76feaef801821f3dc18358b821dc684cf