CHROMIUM: clk: rockchip: Add low jitter 428 MHz for better 25.175 HDMI
Same as we've been doing recently.  In this case:
  428000000 / 17. = 25176471

That's very very close to the desired rate.  This is also a very low
jitter source clock.

Note that I'll follow my typical pattern of leaving the "BWADJ" at
default for clocks < 3 GHz, though there's been no difinitive tests that
this rule if thumb is right.  We should easily be low jitter enough,
though.

BUG=chrome-os-partner:49507
TEST=25.175 still works OK and uses new PLL setting.

Change-Id: I1724e59d2aaebbefd9e0369fcf7d2e661626f422
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/327711
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>