CHROMIUM: drm: rockchip/dw_hdmi: Add 83.5, remove 83.52
Same as we've been doing recently.  Now that we're accepting PLL rates
with larger Fvco for HDMI, we can come up with PLL settings that make an
even multiple of 83.5 MHz.  That means we can remove 83.52 and replace
with exactly 83.5 in the list of supported rates.

BUG=chrome-os-partner:50057
TEST=83.5 still works OK and uses new PLL setting.

Change-Id: I5c285cddacb0f9f5405cf0ac4b05d7feb1bfc1ed
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/325514
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
(cherry picked from commit fed246717a62fa3ea1cfa77d33afce9f425726e7)
Reviewed-on: https://chromium-review.googlesource.com/329000