UPSTREAM: mfd: intel_quark_i2c_gpio: Use clkdev_create()
Convert this driver to use clkdev_create() instead of
clk_register_clkdevs(). The latter API is only used by this driver,
although this driver only allocates one clk to add anyway.
Furthermore, this driver allocates the clk_lookup structure with
devm, but clkdev_drop() will free that structure when passed,
leading to a double free when this driver is removed. Clean it
all up and pave the way for the removal of clk_register_clkdevs().

BUG=b:62106805
TEST=Build i386:allmodconfig

Change-Id: Ia77e9bea34ff25e56477d3866207863bdf950d2a
Cc: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Cc: Russell King <linux@arm.linux.org.uk>
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit c4726abce63b)
Reviewed-on: https://chromium-review.googlesource.com/682415
Reviewed-by: Dylan Reid <dgreid@chromium.org>