UPSTREAM: ext4: shortcut setting of xattr to the same value
When someone tried to set xattr to the same value (i.e., not changing
anything) we did all the work of removing original xattr, possibly
breaking references to shared xattr block, inserting new xattr, and
merging xattr blocks again. Since this is not so rare operation and it
is relatively cheap for us to detect this case, check for this and
shortcut xattr setting in that case.

BUG=chromium:766786
TEST=build/boot on caroline, xfstest

Change-Id: Ie7d5da9efd31221f9c7941cd889cf7aeb0c0f0c9
Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
(cherry picked from commit 3fd164629d25b04f291a79a013dcc7ce1a301269)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/677081
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>