UPSTREAM: usb: dwc2: Fix issues in dwc2_complete_non_isoc_xfer_ddma()
Fixes a static analysis issue in dwc2_complete_non_isoc_xfer_ddma(). The
qtd was being passed to a function after being freed. It was not being
used in the function so this doesn't fix any bugs. But it fixes up the
warning and makes the code safer by setting qtd to NULL and not using it
at all.

BUG=chrome-os-partner:50683
TEST=USB still works OK

Change-Id: Ia3537b4b569939fcf49fbb6b200e45b063ea9804
Reported-by: Felipe Balbi <balbi@kernel.org>
Signed-off-by: John Youn <johnyoun@synopsys.com>
Signed-off-by: Felipe Balbi <balbi@kernel.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from git.kernel.org balbi/usb.git next
 commit 7ed28fc6b9ea5cb6f5208b5abc22e75997bb9c83)
Reviewed-on: https://chromium-review.googlesource.com/330154
Reviewed-by: Julius Werner <jwerner@chromium.org>