CHROMIUM: HACK: ASoC: hdac_hdmi: runtime get controller on prepare
If the audio controller for hdac_hdmi is runtime suspended during the
soc-core device's suspend callback, there can be errors while
trying to talk to the hdac_hdmi device.

soc-core also has separate suspend/resume callbacks for the codecs
that are outside the PM subsystem, and the soc-core device is outside
of the device tree for its codecs.

Because of these reasons, runtime resume the audio controller during a
prepare callback for the hdac hdmi device. The hdac hdmi codec
firmware may timeout on a response during runtime resume. This delays
suspend, so don't runtime resume the codecs for now. When that is
resolved, we should revert that patch and call pm_runtime_get_sync on
all of the codecs for a prepare callback in soc-core.

BUG=chrome-os-partner:50532
TEST=suspend_stress_test on chell

Change-Id: Ib7a3f900c53ed11febab695a08b982c24b2e78c1
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/330961
(cherry picked from commit 82e47f0ed0afb623c4f7a1121185edbe3d5ff8af)
Reviewed-on: https://chromium-review.googlesource.com/331800