CHROMIUM: ASoC: hdac_hdmi: make sure device is suspended
The device driver previously relied on the codec to be runtime
suspended for system suspend. This adds a check for a codec suspend
callback which will suspend the codec if it is not runtime suspended.

This also recovers the codec device to its prior state on resume. If
the device was runtime suspended, resume will resuspend the device at
the end of the codec resume.

BUG=chrome-os-partner:50532
TEST=suspend_stress_test on chell

Change-Id: I5d019e233d064fa0ff820780c2a2cf3a7c9d5c4c
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/330960
(cherry picked from commit 9855cf08234ae4c2e9ade901383d9b284a202842)
Reviewed-on: https://chromium-review.googlesource.com/331790