BACKPORT: FROMLIST: ARM: errata: Workaround erratum A12 821420
This erratum has a very simple workaround (set a bit in a register), so
let's apply it.  Apparently the workaround's downside is a very slight
power impact.

Note that applying this errata fixes deadlocks that are easy to
reproduce with real world applications.

The arguments for why this needs to be in the kernel are similar to the
arugments made in the patch "Workaround errata A12 818325/852422 A17
852423".

Conflicts:
	arch/arm/mm/proc-v7.S

BUG=chrome-os-partner:47192
TEST=kernel_CheckArmErrata autotest

Change-Id: Iacb0c57d1ce49e6411eb35a25ac07f66733b584a
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/8548411/)
Reviewed-on: https://chromium-review.googlesource.com/331744
Reviewed-by: Olof Johansson <olofj@chromium.org>