BACKPORT: FROMLIST: ARM: errata: Workaround errata A12 825619 / A17 852421
The workaround for both errata is to set bit 24 in the diagnostic
register.  There are no known end-user bugs solved by fixing this
errata, but the fix is trivial and it seems sane to apply it.

The arguments for why this needs to be in the kernel are similar to the
arugments made in the patch "Workaround errata A12 818325/852422 A17
852423".

Conflicts:
	arch/arm/mm/proc-v7.S

BUG=chrome-os-partner:51002
TEST=kernel_CheckArmErrata autotest

Change-Id: Ib10c9beda34eda3be38af6563267737c979e50ca
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/8548401/)
Reviewed-on: https://chromium-review.googlesource.com/331747
Reviewed-by: Olof Johansson <olofj@chromium.org>
(cherry picked from commit 00aaae2cfc11a165ed089888f429f289c414d501)
Reviewed-on: https://chromium-review.googlesource.com/333090