CHROMIUM: iwl7000: store fw memory segments length and addresses in run-time
Currently reading the fw memory segments is done according to
addresses and data length that are hard-coded.
Lately a new tlv was appended to the ucode, that contains
the data type, length and address.
Parse this tlv, and in run-time store the memory segments length
and addresses that would be dumped upon a fw error.

Signed-off-by: Golan Ben-Ami <golan.ben.ami@intel.com>
iwl7000-tree: f030026c3103ce0551cd6306221ec6fb1303171e