HACK: [media] vb2: Don't unmap a dmabuf while dequeueing
The hardware still needs to read from those buffers as reference
frames while decoding. To solve the issue, a dmabuf cannot be
unmapped when it was dequeued. No memory will leaked since the dmabuf
will still be unmapped and detached when the buffer is put.

BUG=None
TEST=Play YT and ApiDemo on ARC
Signed-off-by: Pawel Osciak <posciak@chromium.org>
Change-Id: I863cade2732971dc9a8e1818e4e3bd8dc1f0ddb7
Reviewed-on: https://chromium-review.googlesource.com/334103
Commit-Ready: Owen Lin <owenlin@chromium.org>
Tested-by: Owen Lin <owenlin@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>