UPSTREAM: mac80211: mesh: handle failed alloc for rmc cache
In the unlikely case that mesh_rmc_init() fails with -ENOMEM,
the rmc pointer will be left as NULL but the interface is still
operational because ieee80211_mesh_init_sdata() is not allowed
to fail.

If this happens, we would blindly dereference rmc when checking
whether a multicast frame is in the cache.  Instead just drop the
frames in the forwarding path.

Signed-off-by: Bob Copeland <me@bobcopeland.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>

(cherry-picked from mac80211 0aa7fabbd5d9da1f8a8fdc3e2837c532bcfa5664)

BUG=b:29897970
TEST=manually on whirlwinds

Change-Id: I74c375f260e6999707f70838cf8524e1122d2a22
Signed-off-by: Julan Hsu <julanhsu@google.com>
Reviewed-on: https://chromium-review.googlesource.com/365200
Reviewed-by: Srinivasa duvvuri <sduvvuri@chromium.org>
Reviewed-by: Zhihong Yu <zhihongyu@chromium.org>