UPSTREAM: tracing: Fix trace_printk() to print when not using bprintk()
The trace_printk() code will allocate extra buffers if the compile detects
that a trace_printk() is used. To do this, the format of the trace_printk()
is saved to the __trace_printk_fmt section, and if that section is bigger
than zero, the buffers are allocated (along with a message that this has
happened).

If trace_printk() uses a format that is not a constant, and thus something
not guaranteed to be around when the print happens, the compiler optimizes
the fmt out, as it is not used, and the __trace_printk_fmt section is not
filled. This means the kernel will not allocate the special buffers needed
for the trace_printk() and the trace_printk() will not write anything to the
tracing buffer.

Adding a "__used" to the variable in the __trace_printk_fmt section will
keep it around, even though it is set to NULL. This will keep the string
from being printed in the debugfs/tracing/printk_formats section as it is
not needed.

BUG=chromium:724788
TEST=Build and run

Change-Id: Ib3174b493a43e4497c3402d5b84a80ce4c991bbf
Reported-by: Vlastimil Babka <vbabka@suse.cz>
Fixes: 07d777fe8c398 "tracing: Add percpu buffers for trace_printk()"
Cc: stable@vger.kernel.org # v3.5+
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 3debb0a9ddb1)
Reviewed-on: https://chromium-review.googlesource.com/510808
Reviewed-by: Dylan Reid <dgreid@chromium.org>
(cherry picked from commit de69bbee2e72f025d344f4af63a6cffee8e79014)
Reviewed-on: https://chromium-review.googlesource.com/513108