UPSTREAM: drm/i915: Make modeset state verifier take crtc as argument.
This will make it easier to keep the crtc checker when atomic
commit is reworked for asynchronous commits. This prevents checking
crtc's that were not part of the state. It's safe to verify disabled
encoders, connectors and dpll's that are not part of the state,
because during modeset connection_mutex is held.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/1458741487-23801-2-git-send-email-maarten.lankhorst@linux.intel.com
Reviewed-by: Matt Roper <matthew.d.roper@intel.com>
[mlankhorst: Extend commit message and rename check to verify.]

(cherry picked from commit e7c84544757a405e430779e2ffe13ebe9b048361)
Signed-off-by: Radhakrishna Sripada <radhakrishna.sripada@intel.com>

BUG=chrome-os-partner:53839
TEST=External monitors connected via MST hub or daisy-chain over DP should come up.

Change-Id: I99b79ae79102649564305752b49384f577782166
Reviewed-on: https://chromium-review.googlesource.com/421893
Commit-Ready: Brian J Lovin <brian.j.lovin@intel.com>
Tested-by: Chenglei Ren <chenglei.ren@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>