CHROMIUM: iwl7000: mac80211: fix RX u64 stats consistency on 32-bit platforms
On 32-bit platforms, the 64-bit counters we keep need to be protected
to be consistently read. Use the u64_stats_sync mechanism to do that.

In order to not end up with overly long lines, refactor the tidstats
assignments a bit.

[again, we don't seem to care much, so let's not worry about bugfix]

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 150dd7edbeb28538c662b4b18febc66612ebac03