CHROMIUM: Instrument driver to better show problems gathering channel load.
These changes are being submitted in order to get them to beta
channel to help diagnose where the system is breaking when
we fail to get channel survey results on one of the main
radios.  It triggers so infrequently that we need a large
test group.  Once triggered, it does not affect traffic on the
device, but it does not seem to self-resolve and it prevents
Waveguide from making channel change recommendations.

This CL does not change any functionality. It merely adds logging.

This CL will be reverted prior to public release.

TEST=Manually verify the modified log messages.
BUG=b:27929407,b:27787219
Signed-off-by: walker <walker@chromium.org>

Change-Id: I804a0fa4a8066eb953936715083e87883a0b798a
Reviewed-on: https://chromium-review.googlesource.com/336451
Commit-Ready: John Walker <walker@google.com>
Tested-by: John Walker <walker@google.com>
Reviewed-by: Srinivasa duvvuri <sduvvuri@chromium.org>