CHROMIUM: Clean up logs added to diagnose B27787219.
During work to help diagnose b/27929407 and b/27787219, additional
logging was added to ap-monitor. Some of this will be generally
useful in the future. Some is either too verbose or very
specific to the diagnosis of that bug. This CL removes the extra
logs and also cleans up the log messages that are generally
useful.

BUG=b:27929407,b:27787219
TEST=Manually verified log messages.
Signed-off-by: walker <walker@chromium.org>

Change-Id: I65e028b46a46e97da939317b5d3ed41d430b48a5
Reviewed-on: https://chromium-review.googlesource.com/337981
Commit-Ready: John Walker <walker@google.com>
Tested-by: John Walker <walker@google.com>
Reviewed-by: Srinivasa duvvuri <sduvvuri@chromium.org>
Reviewed-by: John Walker <walker@google.com>